mini-projects/hf-inference/pyproject.toml
[project]
name = "hf-inference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.29.3",
    "python-dotenv>=1.0.1",
]
